presque

+++ ft_printf.h
@@ -13,4 +13,4 @@ int		ft_printf(const char *str, ...);
 int	    ft_putnbr(int nb, int count);
 int	    ft_putunsignednbr(unsigned int nb, int count);
 int	    ft_putchar(int c);
-int     ft_puthexnbr(int nb, int minmaj, int count);
+int     ft_puthexnbr(size_t nb, int minmaj, int count);
